requirements.txt
lmdb
tqdm
torch
torchvision
Pillow
requests
matplotlib
imageio
imageio-ffmpeg
albumentations
imgaug
tensorboard
Ninja
#setuptools==59.5.0